Capstone-1/requirements.txt
streamlit==1.31.0
openai>=2.17.0
pandas==2.2.0
plotly==5.18.0
PyGithub==2.1.1
python-dotenv==1.0.1
